use settings
Source: https://github.com/msupply-foundation/open-msupply/commit/1f2a1f9e68631bac01ced32232bd8d1d9d1fc39f

+++ server/graphql/general/src/mutations/label_printer_settings.rs
@@ -10,7 +10,7 @@ pub struct LabelPrinterSettingsInput {
     pub address: String,
     pub label_height: i32,
     pub label_width: i32,
-    pub port: i32,
+    pub port: u16,
 }
 
 #[derive(SimpleObject)]
